apps/api/pyproject.toml
[project]
name = "takeout-api"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "python-slugify>=8.0.4",
]

[build-system]
requires = ["setuptools>=69.0"]
build-backend = "setuptools.build_meta"
